Added "Remove from groups" action on table toolbar
