+++ frontend/app/src/entities/nodes/object/ui/object-table/toolbar/actions/groups/group-item.tsx
