+++ frontend/app/src/entities/nodes/object/ui/object-table/toolbar/actions/groups/processing-groups-panel.tsx
@@ -1,31 +1,34 @@
-import { ProcessingGroupItem } from "@/entities/nodes/object/ui/object-table/toolbar/add-to-groups/processing-group-item";
 import {
   GroupPanelBody,
   GroupPanelFooter,
   GroupPanelHeader,
-} from "@/entities/nodes/object/ui/object-table/toolbar/group-panel";
+} from "@/entities/nodes/object/ui/object-table/toolbar/actions/groups/group-panel";
+import {
+  ProcessingGroupItem,
+  ProcessingGroupItemProps,
+} from "@/entities/nodes/object/ui/object-table/toolbar/actions/groups/processing-group-item";
 import { RelationshipNode } from "@/entities/nodes/relationships/domain/types";
-import { NodeObject } from "@/entities/nodes/types";
 import { Button } from "@/shared/components/buttons/button-primitive";
 import { pluralize } from "@/shared/utils/string";
 import React from "react";
 import { ListBox } from "react-aria-components";
 
-export interface ProcessingGroupsPanelProps {
+export interface ProcessingGroupsPanelProps extends Omit<ProcessingGroupItemProps, "group"> {
   selectedGroups: RelationshipNode[];
-  selectedRows: NodeObject[];
-  onSuccess: () => void;
 }
 
 export function ProcessingGroupsPanel({
   selectedGroups,
-  selectedRows,
+  mutationFn,
   onSuccess,
 }: ProcessingGroupsPanelProps) {
   const [successCount, setSuccessCount] = React.useState(0);
 
   return (
-    <div className="border-l min-w-[15rem] max-w-sm max-h-[12rem] flex flex-col">
+    <div
+      className="border-l min-w-[15rem] max-w-sm max-h-[12rem] flex flex-col"
+      data-testid="processing-groups-panel"
+    >
       <GroupPanelHeader>
         {successCount} / {pluralize(selectedGroups.length, "group")} updated successfully
       </GroupPanelHeader>
@@ -38,10 +41,10 @@ export function ProcessingGroupsPanel({
         >
           {(group) => (
             <ProcessingGroupItem
-              group={group}
-              selectedRows={selectedRows}
-              onSuccess={() => setSuccessCount((prev) => prev + 1)}
               key={group.id}
+              group={group}
+              mutationFn={mutationFn}
+              onSuccess={() => setSuccessCount((prev) => prev + 1)}
             />
           )}
         </ListBox>
