+++ frontend/app/src/entities/nodes/object/ui/object-table/toolbar/actions/groups/toolbar-add-to-groups-action.tsx
@@ -1,0 +1,41 @@
+import { useCurrentBranch } from "@/entities/branches/ui/branches-provider";
+import { BulkMutateGroups } from "@/entities/nodes/object/ui/object-table/toolbar/actions/groups/bulk-mutate-groups";
+import { ToolbarButton } from "@/entities/nodes/object/ui/object-table/toolbar/toolbar-button";
+import { addRelationships } from "@/entities/nodes/relationships/domain/add-relationships/add-relationships";
+import { NodeObject } from "@/entities/nodes/types";
+import { Popover, PopoverDialog } from "@/shared/components/aria/popover";
+import { DialogTrigger } from "react-aria-components";
+
+export interface ToolbarAddToGroupActionProps {
+  selectedRows: Array<NodeObject>;
+}
+
+export function ToolbarAddToGroupsAction({ selectedRows }: ToolbarAddToGroupActionProps) {
+  const { currentBranch } = useCurrentBranch();
+
+  return (
+    <DialogTrigger>
+      <ToolbarButton>Add to groups</ToolbarButton>
+
+      <Popover placement="top start">
+        <PopoverDialog className="p-0">
+          {({ close }) => {
+            return (
+              <BulkMutateGroups
+                mutationFn={async (group) => {
+                  await addRelationships({
+                    objectId: group.id,
+                    relationshipName: "members",
+                    relationshipIds: selectedRows.map((row) => row.id),
+                    branchName: currentBranch.name,
+                  });
+                }}
+                onSuccess={close}
+              />
+            );
+          }}
+        </PopoverDialog>
+      </Popover>
+    </DialogTrigger>
+  );
+}
