+++ frontend/app/src/entities/nodes/object/ui/object-table/toolbar/object-table-toolbar.tsx
@@ -1,4 +1,5 @@
-import { ToolbarAddToGroupsAction } from "@/entities/nodes/object/ui/object-table/toolbar/add-to-groups/toolbar-add-to-groups-action";
+import { ToolbarAddToGroupsAction } from "@/entities/nodes/object/ui/object-table/toolbar/actions/groups/toolbar-add-to-groups-action";
+import { ToolBarRemoveFromGroupsAction } from "@/entities/nodes/object/ui/object-table/toolbar/actions/groups/toolbar-remove-from-groups-action";
 import { NodeObject } from "@/entities/nodes/types";
 import { classNames } from "@/shared/utils/common";
 import { XIcon } from "lucide-react";
@@ -15,7 +16,7 @@ export function ObjectTableToolbar({ selectedRows, onClose }: ObjectTableSelecti
       isOpen
       className={classNames(
         "fixed bottom-10 left-1/2 -translate-x-1/2 z-50",
-        "text backdrop-blur-lg p-2 shadow-lg rounded-2xl border border-neutral-300",
+        "text backdrop-blur-lg p-2 shadow-lg rounded-2xl border border-neutral-300 text-sm",
         "data-[entering]:animate-in data-[entering]:fade-in-0 data-[entering]:zoom-in-95 data-[entering]:slide-in-from-left-1/2",
         "data-[exiting]:duration-300 data-[exiting]:animate-out data-[exiting]:fade-out-0 data-[exiting]:zoom-out-95 data-[exiting]:slide-out-to-left-1/2",
         "flex items-center gap-2 outline-none"
@@ -30,6 +31,7 @@ export function ObjectTableToolbar({ selectedRows, onClose }: ObjectTableSelecti
       </AriaButton>
 
       <ToolbarAddToGroupsAction selectedRows={selectedRows} />
+      <ToolBarRemoveFromGroupsAction selectedRows={selectedRows} />
     </AriaModal>
   );
 }
